Adds ReceiptService and implements some of its logic handling

+++ Services/Interfaces/IAccountService.cs
@@ -3,7 +3,7 @@ using PucBank.Models;
 namespace PucBank.Services.Interfaces;
 public interface IAccountService
 {
-    Account CreateAccount(string firstName, string lastName, int balance);
-    void Deposit(Account user, int depositAmount);
-    void Withdraw(Account user, int withdrawAmount);
+    Account CreateAccount(string firstName, string lastName, double balance);
+    void Deposit(Account user, double depositAmount);
+    void Withdraw(Account user, double withdrawAmount);
 }
